+++ webapi/src/main/scala/org/knora/webapi/responders/v2/ResourcesResponderV2.scala
@@ -27,12 +27,9 @@ import org.knora.webapi.core.MessageHandler
 import org.knora.webapi.core.MessageRelay
 import org.knora.webapi.messages.*
 import org.knora.webapi.messages.IriConversions.*
-import org.knora.webapi.messages.store.sipimessages.SipiGetTextFileRequest
-import org.knora.webapi.messages.store.sipimessages.SipiGetTextFileResponse
 import org.knora.webapi.messages.util.*
 import org.knora.webapi.messages.util.rdf.*
 import org.knora.webapi.messages.util.search.gravsearch.GravsearchParser
-import org.knora.webapi.messages.util.standoff.StandoffTagUtilV2
 import org.knora.webapi.messages.v2.responder.CanDoResponseV2
 import org.knora.webapi.messages.v2.responder.SuccessResponseV2
 import org.knora.webapi.messages.v2.responder.resourcemessages.*
@@ -44,56 +41,45 @@ import org.knora.webapi.messages.v2.responder.valuemessages.*
 import org.knora.webapi.responders.IriLocker
 import org.knora.webapi.responders.IriService
 import org.knora.webapi.responders.Responder
-import org.knora.webapi.responders.admin.PermissionsResponder
 import org.knora.webapi.responders.v2.resources.CreateResourceV2Handler
 import org.knora.webapi.slice.admin.domain.model.KnoraProject.ProjectIri
 import org.knora.webapi.slice.admin.domain.model.Permission
 import org.knora.webapi.slice.admin.domain.model.User
 import org.knora.webapi.slice.admin.domain.model.UserIri
 import org.knora.webapi.slice.admin.domain.service.KnoraProjectService
-import org.knora.webapi.slice.admin.domain.service.LegalInfoService
 import org.knora.webapi.slice.api.v2.GraphDirection
 import org.knora.webapi.slice.api.v2.VersionDate
 import org.knora.webapi.slice.api.v2.ontologies.LastModificationDate
 import org.knora.webapi.slice.common.ResourceIri
 import org.knora.webapi.slice.common.repo.rdf.Vocabulary.KnoraBase as KB
 import org.knora.webapi.slice.common.service.IriConverter
-import org.knora.webapi.slice.ontology.domain.service.OntologyRepo
 import org.knora.webapi.slice.resources.repo.ChangeResourceMetadataQuery
 import org.knora.webapi.slice.resources.repo.DeleteResourceQuery
 import org.knora.webapi.slice.resources.repo.EraseResourceQuery
 import org.knora.webapi.slice.resources.repo.GetAllResourcesInProjectPrequery
 import org.knora.webapi.slice.resources.repo.GetGraphDataQuery
 import org.knora.webapi.slice.resources.repo.GetResourceValueVersionHistoryQuery
-import org.knora.webapi.slice.resources.repo.service.ResourcesRepo
 import org.knora.webapi.slice.resources.service.ReadResourcesService
-import org.knora.webapi.slice.resources.service.ValueContentValidator
 import org.knora.webapi.slice.search.repo.GetIncomingImageLinksGravsearchQuery
+import org.knora.webapi.store.iiif.api.SipiService
 import org.knora.webapi.store.iiif.errors.SipiException
 import org.knora.webapi.store.triplestore.api.TriplestoreService
 import org.knora.webapi.store.triplestore.api.TriplestoreService.Queries.Select
 import org.knora.webapi.store.triplestore.api.TriplestoreService.Queries.Update
 import org.knora.webapi.util.FileUtil
 
-final case class ResourcesResponderV2(
-  private val appConfig: AppConfig,
-  private val constructResponseUtilV2: ConstructResponseUtilV2,
-  private val iriConverter: IriConverter,
-  private val iriService: IriService,
-  private val legalInfoService: LegalInfoService,
-  private val messageRelay: MessageRelay,
-  private val ontologyRepo: OntologyRepo,
-  private val permissionUtilADM: PermissionUtilADM,
-  private val permissionsResponder: PermissionsResponder,
-  private val projectService: KnoraProjectService,
-  private val resourceUtilV2: ResourceUtilV2,
-  private val resourcesRepo: ResourcesRepo,
-  private val searchResponderV2: SearchResponderV2,
-  private val standoffTagUtilV2: StandoffTagUtilV2,
-  private val triplestore: TriplestoreService,
-  private val valueValidator: ValueContentValidator,
-  private val readResources: ReadResourcesService,
-  private val createHandler: CreateResourceV2Handler,
+final class ResourcesResponderV2(
+  appConfig: AppConfig,
+  iriConverter: IriConverter,
+  iriService: IriService,
+  messageRelay: MessageRelay,
+  projectService: KnoraProjectService,
+  resourceUtilV2: ResourceUtilV2,
+  searchResponderV2: SearchResponderV2,
+  sipiService: SipiService,
+  triplestore: TriplestoreService,
+  readResources: ReadResourcesService,
+  createHandler: CreateResourceV2Handler,
 )(implicit val stringFormatter: StringFormatter)
     extends MessageHandler {
 
@@ -573,16 +559,7 @@ final case class ResourcesResponderV2(
 
     for {
       gravsearchTemplateUrl <- recoveredGravsearchUrlTask
-      response              <- messageRelay
-                    .ask[SipiGetTextFileResponse](
-                      SipiGetTextFileRequest(
-                        fileUrl = gravsearchTemplateUrl,
-                        requestingUser = KnoraSystemInstances.Users.SystemUser,
-                        senderName = this.getClass.getName,
-                      ),
-                    )
-      gravsearchTemplate: String = response.content
-
+      gravsearchTemplate    <- sipiService.getTextFileRequest(gravsearchTemplateUrl, this.getClass.getName)
     } yield gravsearchTemplate
 
   }
@@ -1288,7 +1265,7 @@ final case class ResourcesResponderV2(
                       Seq(
                         JsonLDObject(
                           Map(
-                            "id"         -> JsonLDString(imageValue.valueIri),
+                            "id"         -> JsonLDString(imageValue.valueIri.value),
                             "type"       -> JsonLDString("Annotation"),
                             "motivation" -> JsonLDString("painting"),
                             "body"       -> JsonLDObject(
@@ -1791,41 +1768,27 @@ object ResourcesResponderV2 {
   val layer = ZLayer.fromZIO {
     for {
       appConfig               <- ZIO.service[AppConfig]
-      constructResponseUtilV2 <- ZIO.service[ConstructResponseUtilV2]
       iriConverter            <- ZIO.service[IriConverter]
       iriService              <- ZIO.service[IriService]
-      legalInfoService        <- ZIO.service[LegalInfoService]
       messageRelay            <- ZIO.service[MessageRelay]
-      ontologyRepo            <- ZIO.service[OntologyRepo]
-      permissionUtilADM       <- ZIO.service[PermissionUtilADM]
-      permissionsResponder    <- ZIO.service[PermissionsResponder]
       projectService          <- ZIO.service[KnoraProjectService]
       resourceUtilV2          <- ZIO.service[ResourceUtilV2]
-      resourcesRepo           <- ZIO.service[ResourcesRepo]
       searchResponderV2       <- ZIO.service[SearchResponderV2]
-      standoffTagUtilV2       <- ZIO.service[StandoffTagUtilV2]
+      sipiService             <- ZIO.service[SipiService]
       stringFormatter         <- ZIO.service[StringFormatter]
       triplestoreService      <- ZIO.service[TriplestoreService]
-      valueContentValidator   <- ZIO.service[ValueContentValidator]
       readResources           <- ZIO.service[ReadResourcesService]
       createResourceV2Handler <- ZIO.service[CreateResourceV2Handler]
       responder                = new ResourcesResponderV2(
                     appConfig,
-                    constructResponseUtilV2,
                     iriConverter,
                     iriService,
-                    legalInfoService,
                     messageRelay,
-                    ontologyRepo,
-                    permissionUtilADM,
-                    permissionsResponder,
                     projectService,
                     resourceUtilV2,
-                    resourcesRepo,
                     searchResponderV2,
-                    standoffTagUtilV2,
+                    sipiService,
                     triplestoreService,
-                    valueContentValidator,
                     readResources,
                     createResourceV2Handler,
                   )(stringFormatter)